+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 #### ⚙️ Internal
 
+- Updated Rust to v1.92.0.
 - Updated dependencies.
 
 ## 0.54.0